Move all wildcards to files in the new wildcards folder This allows for easy and intuitive grouping.
It also breaks everything.
