+++ desktop/server/wildcards/_template.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const helpers = require('../core/helpers')
+
+const MODULENAME = {
+  // Functions go here
+}
+
+module.exports = MODULENAME
